chore: fix all the cycle dependency issues
