+++ src/helpers/APIHelper.js
@@ -2,10 +2,12 @@ import axios from 'axios';
 import * as Sentry from '@sentry/react-native';
 
 import { API_URL } from '../constants/url';
-import I18n from '../i18n';
+import I18n from '../i18n'; 
 
 import { showToast } from './ToastHelper';
-import { getHeaders, getBaseUrl, handleLogout } from '../services/auth';
+import { getHeaders, getBaseUrl } from '../services/auth';
+import { handleLogout } from '../reducer/authHelper'; 
+import { getStore } from '../reducer/storeAccessor';
 
 const parseErrorCode = error => {
   Sentry.captureException(error);
@@ -24,7 +26,7 @@ const API = axios.create();
 // Request parsing interceptor
 API.interceptors.request.use(
   async config => {
-    const headers = await getHeaders() 
+    const headers = await getHeaders();
     config.baseURL = await getBaseUrl();
     const configHeaders = config.headers;
     if (headers) {
