+++ src/reducer/authHelper.js
@@ -1,0 +1,6 @@
+import { getStore } from '../reducer/storeAccessor';
+
+export const handleLogout = async () => {
+  const store = await getStore();
+  store.dispatch({ type: 'auth/logout' });
+};
