+++ src/reducer/conversationSlice.js
@@ -2,6 +2,7 @@ import { createSlice, createEntityAdapter } from '@reduxjs/toolkit';
 import actions from './conversationSlice.action';
 import { MESSAGE_TYPES } from 'constants';
 import { findPendingMessageIndex } from '../helpers/conversationHelpers';
+import { addOrUpdateMessageHelper } from './conversationHelpers';
 export const conversationAdapter = createEntityAdapter({
   selectId: conversation => conversation.id,
 });
@@ -70,31 +71,7 @@ const conversationSlice = createSlice({
       }
     },
     addOrUpdateMessage: (state, action) => {
-      const message = action.payload;
-
-      const { conversation_id: conversationId } = message;
-      if (!conversationId) {
-        return;
-      }
-      const conversation = state.entities[conversationId];
-
-      // If the conversation is not present in the store, we don't need to add the message
-      if (!conversation) {
-        return;
-      }
-      // If the message type is incoming, set the can reply to true
-      if (message.message_type === MESSAGE_TYPES.INCOMING) {
-        conversation.can_reply = true;
-      }
-      const pendingMessageIndex = findPendingMessageIndex(conversation, message);
-      if (pendingMessageIndex !== -1) {
-        conversation.messages[pendingMessageIndex] = message;
-      } else {
-        conversation.messages.push(message);
-        conversation.timestamp = message.created_at;
-        const { conversation: { unread_count: unreadCount = 0 } = {} } = message;
-        conversation.unread_count = unreadCount;
-      }
+      addOrUpdateMessageHelper(state, action.payload);
     },
     updateConversationLastActivity: (state, action) => {
       const { conversationId, lastActivityAt } = action.payload;
